+++ fighters/common/src/function_hooks/vtables/ganon.rs
