+++ fighters/common/src/function_hooks/vtables/lucario.rs
