+++ fighters/common/src/function_hooks/vtables/mod.rs
@@ -1,7 +1,21 @@
 use super::*;
 
+mod donkey;
+mod ganon;
+mod lucario;
+mod ptrainer;
+mod littlemac;
 mod rockman;
+mod krool;
+mod brave;
 
 pub fn install() {
+    donkey::install();
+    ganon::install();
+    lucario::install();
+    ptrainer::install();
+    littlemac::install();
     rockman::install();
+    krool::install();
+    brave::install();
 }
